retail_agency/requirements.txt
agency-swarm>=0.1.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
notion-client>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # For PostgreSQL
openai>=1.0.0
slack-bolt>=1.18.0
pydantic>=2.8.2
pytz>=2024.1
cloud-sql-python-connector[pg8000]>=1.16.0
google-auth>=2.38.0
google-cloud-core>=2.0.0
langchain-openai>=0.0.5  # For OpenAI integration
pg8000>=1.30.5  # Python PostgreSQL driver 